// src/data/card-crop-selections.generated.ts
export const cardCropSelectionsGenerated: Readonly<Record<string, CardCropRecord>> = {
  "43935854_p0_master1200.webp": {
    "filename": "43935854_p0_master1200.webp",
    "fit": "cover",
    "preferredFrame": "diagonal-left",
    "schemaVersion": 2,
    "selection": "diagonal-left",
    "transforms": {
      "diagonal-left": {
        "x": 60.78125,
        "y": 28.531624407913057,
        "zoom": 1
      },
      "diagonal-right": {
        "x": 100,
        "y": 32.14264913238833,
        "zoom": 1
      }
    },
    "cropRects": {
      "diagonal-left": {
        "height": 1,
        "width": 0.892857142857143,
        "x": 0.06512276785714279,
        "y": 0
      },
      "diagonal-right": {
        "height": 1,
        "width": 0.892857142857143,
        "x": 0.10714285714285704,
        "y": 0
      }
    },
    "updatedAt": "2026-08-31T17:25:31.931Z"
  },
// ...
}